+++ src/nvlib/gui/pop_up/backup_options_ctrl.py
@@ -5,11 +5,13 @@ For further information see https://github.com/peter88213/novelibre
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 import os
-from pathlib import Path
 from tkinter import filedialog
 
 from mvclib.controller.sub_controller import SubController
+from nvlib.controller.services.nv_help import NvHelp
+from nvlib.model.file.doc_open import open_document
 from nvlib.novx_globals import norm_path
+from nvlib.nv_globals import HOME_DIR
 from nvlib.nv_globals import prefs
 from nvlib.nv_locale import _
 
@@ -31,28 +33,18 @@ class BackupOptionsCtrl(SubController):
             self._ui.set_status(f'#{_("Backup directory not found")}. {_("Please check the setting")}.')
             return
 
-        try:
-            os.startfile(norm_path(prefs['backup_dir']))
-            # Windows
-        except:
-            try:
-                os.system('xdg-open "%s"' % norm_path(prefs['backup_dir']))
-                # Linux
-            except:
-                try:
-                    os.system('open "%s"' % norm_path(prefs['backup_dir']))
-                    # Mac
-                except:
-                    pass
-        return 'break'
+        open_document(prefs['backup_dir'])
+
+    def open_help(self, event=None):
+        NvHelp.open_help_page(f'tools_menu.html#{_("Backup options").lower()}')
 
     def set_backup_dir(self):
         self._ui.restore_status()
         if os.path.isdir(prefs['backup_dir']):
-            startDir = prefs['backup_dir']
+            initDir = prefs['backup_dir']
         else:
-            startDir = str(Path.home()).replace('\\', '/')
-        backupDir = filedialog.askdirectory(initialdir=startDir)
+            initDir = HOME_DIR
+        backupDir = filedialog.askdirectory(initialdir=initDir)
         if not backupDir:
             self._ui.set_status(f'#{_("Action canceled by user")}.')
             return